+++ jwst/imprint/tests/test_imprint.py
@@ -12,7 +12,9 @@ def test_step(make_imagemodel):
     """Assert that the results should be all zeros.
     """
     im = make_imagemodel(10, 10)
-    result = ImprintStep.call(im, im)
+    imprint = []
+    imprint.append(im)
+    result = ImprintStep.call(im, imprint)
 
     assert result.meta.cal_step.imprint == 'COMPLETE'
     assert result.data.sum() == 0