+++ jwst/pipeline/calwebb_spec2.py
@@ -209,7 +209,8 @@ class Spec2Pipeline(Pipeline):
 
             # Decide on what steps can actually be accomplished based on the
             # provided input.
-            self._step_verification(exp_type, members_by_type, multi_int)
+
+            self._step_verification(exp_type, science, members_by_type, multi_int)
 
             # Start processing the individual steps.
             # `assign_wcs` is the critical step. Without it, processing
@@ -237,8 +238,21 @@ class Spec2Pipeline(Pipeline):
                         raise RuntimeError('Cannot determine WCS.')
 
         # Steps whose order is the same for all types of input.
-        calibrated = self.bkg_subtract(calibrated, members_by_type['background'])
+
+        # Leakcal subtraction (imprint)  occurs before background subtraction on a per-exposure basis.
+        # If there is only one `imprint` member, this imprint exposure is subtracted from all the
+        # science and background exposures.  Otherwise, there will be as many `imprint` members as
+        # there are science plus background members.
+
         calibrated = self.imprint_subtract(calibrated, members_by_type['imprint'])
+
+        # for each background image subtract an associated leak cal
+        for i, bkg_file in enumerate(members_by_type['background']):
+            bkg_imprint_sub = self.imprint_subtract(bkg_file, members_by_type['imprint'])
+            members_by_type['background'][i] = bkg_imprint_sub
+
+        calibrated = self.bkg_subtract(calibrated, members_by_type['background'])
+
         calibrated = self.msa_flagging(calibrated)
 
         # The order of the next few steps is tricky, depending on mode:
@@ -320,7 +334,7 @@ class Spec2Pipeline(Pipeline):
 
         return calibrated
 
-    def _step_verification(self, exp_type, members_by_type, multi_int):
+    def _step_verification(self, exp_type, input, members_by_type, multi_int):
         """Verify whether requested steps can operate on the given data
 
         Though ideally this would all be controlled through the pipeline
@@ -350,14 +364,15 @@ class Spec2Pipeline(Pipeline):
                 self.log.debug('Science data does not allow direct background subtraction. Skipping "bkg_subtract".')
                 self.bkg_subtract.skip = True
 
-        # Check for imprint subtraction
+        # Check for imprint subtraction. If we have a background then we could have an imprint image associated with the
+        # background.
         imprint = members_by_type['imprint']
         if not self.imprint_subtract.skip:
             if len(imprint) > 0 and (exp_type in ['NRS_MSASPEC', 'NRS_IFU']
                                      or is_nrs_ifu_flatlamp(input)):
-                if len(imprint) > 1:
+                if len(imprint) > 1 and (exp_type in ['NRS_MSASPEC'] or is_nrs_ifu_flatlamp(input)):
                     self.log.warning('Wrong number of imprint members')
-                members_by_type['imprint'] = imprint[0]
+                    members_by_type['imprint'] = imprint[0]
             else:
                 self.log.debug('Science data does not allow imprint processing. Skipping "imprint_subtraction".')
                 self.imprint_subtract.skip = True